cleanup event system pretends explosion event

+++ BukkitPlayground/src/ch/bukkit/playground/util/DateHelper.java
@@ -8,18 +8,18 @@ import java.util.Date;
 public class DateHelper {
 
     public static String format(Date date) {
-        if(date == null) {
+        if (date == null) {
             return "<unknown>";
         }
         SimpleDateFormat dateFormat = new SimpleDateFormat("HH:mm:ss");
         return dateFormat.format(date);
     }
 
-    public static long getMillisForMinutes(int minutes) {
+    public static long getMillisForMinutes(long minutes) {
         return minutes * 60000;
     }
 
-    public static BigDecimal getMinutesForMillis(int millis) {
+    public static BigDecimal getMinutesForMillis(long millis) {
         return new BigDecimal(millis / 60000.).setScale(1, RoundingMode.HALF_UP);
     }
 }
